Improve handling of incoming i64 syscall arguments.
Add new `receiveI64ParamAsI53` to replace the old
`receiveI64ParamAsDouble`.  The new function checks the value fits
within I53 and requires an `onError` value to be returned in the case of
out-of-bounds.

`receiveI64ParamAsI53` makes use of the existing `convertI32PairToI53`
utility.

There is now some amount of duplication between `test_parseTools`
and `js_library_i64_params` which I would prefer to address as
a followup.